Fixed bugs from the review. Updated methods in repositories "Get by id" after adding join to CRUD

+++ CoCowork.DB/Tables/ProductOrder.sql
@@ -4,7 +4,7 @@
     [ProductId] INT NOT NULL, 
     [OrderId] INT NOT NULL, 
     [Amount] INT NOT NULL, 
-    [SubtotalPrice] INT NOT NULL,
+    [SubtotalPrice] DECIMAL(10, 2) NOT NULL,
     CONSTRAINT [FK_ProductOrder_ToProduct] FOREIGN KEY ([ProductId]) REFERENCES [dbo].[Product]([Id]),
     CONSTRAINT [FK_OrderId_ToOrder] FOREIGN KEY ([OrderId]) REFERENCES [dbo].[Order]([Id]),
